Use primary creds for heat
If the account file has multiple accounts (e.g. 1 normal user and 1
admin), using accounts.hash_dict['creds'] pulls in all the creds. The
code then loops through all the accounts eventually using the last one.
This might not be a normal user cred but an admin cred, which is not
what we want.

Change to use get_primary_creds() which ensures that this is a normal
user cred.

TODO: we should set the admin creds here too, e.g.
self.conf.admin_username

Change-Id: I52e80e6ce19fcd6e4dd16142bcd8315cba784158